pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "favicon-generator"
version = "0.1.0"
description = "A CLI tool to generate .ico favicon files from images"
readme = "README.md"
authors = [
    {name = "Mike Horrocks"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "Pillow>=10.0.0",
    "click>=8.0.0",
]

[project.scripts]
favicon-generator = "favicon_generator.cli:main"

[project.urls]
Homepage = "https://github.com/mikeh74/favicon-generator"
Repository = "https://github.com/mikeh74/favicon-generator"

[tool.setuptools.packages.find]
where = ["."]
include = ["favicon_generator*"]
